requirements.txt
faiss-cpu==1.7.4
uuid==1.30
aiohttp
requests>=2.20
tqdm
PyYAML>=5.3
async-timeout<5.0.0,>=4.0.0
tenacity<9.0.0,>=8.1.0
numpy<2,>=1
SQLAlchemy<3,>=1.4
langsmith<0.2.0,>=0.1.17
pydantic<3,>=1
jsonpatch<2.0,>=1.33
packaging<24.0,>=23.2
tiktoken<1,>=0.7
openai
faiss-cpu==1.7.4
langchain==0.2.3
langchain-core
langchain-openai
langchain-text-splitters
langchain-community
python-dotenv